+++ src/ruby/genotyping-workflows/test/test_zcall_workflow.rb
@@ -1,6 +1,6 @@
 #-- encoding: UTF-8
 #
-# Copyright (c) 2012 Genome Research Ltd. All rights reserved.
+# Copyright (c) 2012, 2016 Genome Research Ltd. All rights reserved.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -58,8 +58,9 @@ class TestWorkflowZCall < Test::Unit::TestCase
 
       FileUtils.copy(File.join(external_data, 'genotyping.db'), dbfile)
       vcf = File.join(external_data, 'sequenom_abvc.vcf')
-      plex_manifest = File.join(external_data,
-                                'W30467_snp_set_info_GRCh37.tsv')
+      plex_0 = File.join(external_data, 'W30467_snp_set_info_GRCh37.tsv')
+      plex_1 = File.join(external_data, 'qc_fluidigm_snp_info_GRCh37.tsv')
+      # plex_1 not needed for workflow, but tests handling multiple plex args
 
       # Only 1 zscore in range; faster but omits threshold evaluation
       # The evaluation is tested by test_zcall_tasks.rb
@@ -70,8 +71,9 @@ class TestWorkflowZCall < Test::Unit::TestCase
                                            :zstart => 6,
                                            :ztotal => 1,
                                            :memory => 2048,
-                                           :vcf => vcf,
-                                           :plex_manifest => plex_manifest }]
+                                           :vcf => [vcf,],
+                                           :plex_manifest => [plex_0, plex_1]
+              }]
       timeout = 1800 # was 720
       log = 'percolate.log'
       result = test_workflow(name, Genotyping::Workflows::GenotypeZCall,